requirements.txt
# Core Dependencies
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0

# Technical Analysis
ta==0.10.2
ta-lib==0.4.28

# Machine Learning (Optional)
scikit-learn==1.2.2
tensorflow==2.12.0

# Visualization and Scheduling
matplotlib==3.7.1
schedule==1.2.0

# Testing and Type Checking
pytest==7.4.0
mypy==1.5.1
flake8==6.1.0

# Performance Optimization
numba==0.57.1

# Logging and Monitoring
loguru==0.7.0

# Async Support
asyncio==3.4.3
aiohttp==3.8.5
